myblog/requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.5
django-cors-headers==4.2.0
django-heroku==0.3.1
django-tinymce==3.6.1
gunicorn==21.2.0
mysqlclient==2.2.0
packaging==23.1
Pillow==10.0.1
psycopg2==2.9.8
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.5.0
